Promising solution! (but currently hardcoded)
So an interesting problem is that the filepaths for a subsheet are
relative to the current sheet... meaning that subsheets in a
subdirectory that refer to each other will use a different path than the
top-level sheet...

I found the bit of code where this goes wrong, and tested it by
hardcoding a check for 'subsheet/' as well, and it works! This is not a
solution, of course, but it does provide 'goal' to work towards.